+++ models/Mprice.go
@@ -3,8 +3,8 @@ package models
 type Price struct {
 	Id        int64  `json:"id"`
 	ProductId int64  `json:"productid"`
-	Mrp       int64  `json:"mrp"`
-	Nop       int64  `json:"nop"`
+	Mrp       float64  `json:"mrp"`
+	Nop       float64  `json:"nop"`
 	Createdon string `json:"createdon"`
 }
